Install mini-css-extract-plugin and split webpack entrypoints into smaller pieces.
When loading pages that don't have a div labeled for loading the React component,
the browser throws a javascript error because the React code is unable to
mount properly. The solution is to not load that code for pages that don't have
any need for an editor React component mount.

In order to achieve this, we need to change the django webpack loader config
so that it doesn't load that javascript code unless we're on thread.html.
Splitting the entrypoints into 'base' and 'editor' helps with this. Base
contains only css so far (but may contain javascript too later on). Editor
contains javascript and css for the React editor component.

Now instead of loading the 'main' bundle (which was the default name), we
load base for all pages and then additionally the editor bundle for the
thread page.

The django bundle loader also has a second parameter to filter bundle
contents by file extension. This is kind of an unnecessary optimization
at this point.