Move DRF endpoints to root conf urls because apparently DRF isn't designed to be nested in an app.
This includes namespacing fixes. So in the path() calls for each DRF router, you can append
an 'app_name' namespace, which is done with a 2-tuple as a parameter to the include() function. This
app_name should be used by internal DRF serializers and stuff to reverse urls. You can also add
an 'instance_name' namespace, which is what comes after the include() call with 'namespace='. This
instance name is used to distinguish different instances of the same router from each other. So,
you can have instance name 'v1' and instance name 'v2'. The internal code of the router still uses
the app_name to reverse urls, but in the rest of the django code or sibling apps, you can use the
instance_names to request a specific version. E.g. v1:post-detail. You can still use the
app_name to reverse a url, and it appears to default to the last instance registered with the url
pattern. So for example, if you have two router instances, 'v1' and 'v2' both with the app name
'api', if you do 'v1:post-detail' you get the v1 instance endpoint. If you do 'api:post-detail'
 you get the v2 instance endpoint. (Assuming that the v2 line is below the v1 line.)