Clean up React hook code.
The old code is just all stuffed into one file. Since we have compiled
the React code as a library that is called through a function in the
html page, we can clean up all the "window.props" stuff.

The first change is to move all the forum submission handling code to
inside the Editor component. Default behavior on form submission
should be to collect the interior of the Window component and populate
the form inside the StatusBar component, then send off an AJAX
request specified by "prop.editor_action".

The only points of customization should be specifying forum post
variables like thread_id, editor action url, and pre/post submit
callbacks. So now the response handling code that inserts a new
row into the HTML page is moved outside the React javascript
library and into an inlined script tag directly on the page. This
is good because only js behavior that is custom to the specific
HTML template is now in the HTML template.

Also if there's time, it might be good to go back and clean up the
names for all the props and how they're threaded into all the
React components.