Add unit and selenium tests for new thread flow. Jesus Christ.
Oh and also change thread-create view to raise a NotImplemented
exception because we need to remember to implement this sometime.

So get this, Django's StaticLiveServerTestCase and
LiveServerTestCase classes are fucking stupid. You're supposed to
do something like this:

class MyTests(StaticLiveServerTestCase):
   @classmethod
   def setUp():
      # stuff

   @classmethod
   def tearDown():
      # stuff

   def test_case1(self):
      # test something

   def test_case2(self):
      # test another thing

So you'd think that work, right??!?? But FUCKING NO. After each test_*
method, the class deletes all the database data. So all the test
cases after the first will fail if they rely on some pre-test setup.
For example, like if you created a Forum object in setUp() and
expected that object to exist across all test cases. You know,
like as suggested by the methods named "setUp" and "tearDown"???

What the fuck.

So there's a comment under the answer here:
https://stackoverflow.com/questions/29378328/django-liveservertestcase-fails-to-load-a-page-when-i-run-multiple-tests

that says you can supposedly create a child class that gives you
the transactional semantics of the TestCase class you'd expect,
but I couldn't get that to work.

The solution was to move the test database set up from the setUp()
method into the beginning of each test_* method.

Fucking shit in my mouth. Godamn.