Change webpack bundle filenaming to distinguish between prod and dev.
This stunning issue, shows how you can replace module.exports assignment
with a lambda function to make the --mode setting visible to the rest
of the configuration:

https://github.com/webpack/webpack/issues/6460

I use this to make the webpack output different filenames depending
on dev or prod. For dev, it prints hashes in the name along with
a ".dev.*" extension. This prevents caching from occuring. For prod,
it prints out the normal filename with no ".dev.*" extension.

I use the ".dev." modifier along with .gitignore to make sure those
files aren't included in the git repo.

And finally, why did I do this?? It's because deploying on heroku
is problematic if I don't include any files in the staticfiles
directory. As specified in the django settings file, the
/backend/assets directory is where all the staticfiles live. However,
currently, there are no files in this directory being tracked by
git. The above changes solve this problem.

So now, before I want to deploy to heroku, I just need to run
"npm run build_prod" inside the /frontend directory and it will
output non-hashed, non-dev names into the assets folder. I then
commit these files into the git repo and do a push.